Resolve issues with invalid names
Invalid names were causing issues with the deletion of objects
from OpenShift. This change fixes it so that all objects are deleted
on deprovisioning of the service. Also there is some (unnecessary...) work
that will 'reverse' the object creation from the order that is
done during provisioning for some components. Turns out it wasn't
really necessary, but it's actually a cleaner way to handle this.