Update nfvha-lab inventory and openshift-ansible version
Note that currently (as of today) the openshift-ansible installation for downstream will result in failure
of the kube-service-catalog because the readiness and liveness prob path of /healthz/ready is invalid. Just
waiting on a spin and deploy of the z-stream where things will match. Just be aware of you're following this
repo and changes as of today. I expect this to be resolved in a couple of days.