A little bit of documentation cleanup
Started doing some of this documentation cleanup, but I'm now realizing
that things are so out of date and no longer really applicable that we're
likely going to need to focus on a complete rewrite of everything.

I'm importing what I've been able to muster up for now, but ultimately
the documentation is going to require some extensive work that I'm not
prepared to tackle currently.