Start moving towards use of ImageStreams
I'm working on migrating to the use of ImageStreams so that spec files
for upstream vs downstream are more consistent and easier to deal with
between environments. It should also allow for offline installs to be
easier to work with.

Unfortunately the AMQ Interconnect Operator doesn't seem to like the
use of ImageStreams right now, so that particular manifest still
points at a direct link to the downstream image for now. An issue has
been opened with the Messaging team to address this in the future.

For now, all changes present allow for the instantiation of SAF using
all components from Red Hat Container Catalog.