Add deploy builder functionality
In order to support multiple environments, we need to dynamically
create some of the manifests used in the deployment of Service
Assurance Framework.

The changes here allow for the spin up of SAF on both a standard
openshift-ansible installed system, and for a simple demo environment
on top of minishift.

We'll do this by providing some simple Ansible that generates a
handful of manifests using some defaults that can be overridden at the
console.

Closes #24