Workaround in deploy.sh for operator race condition
If your environment is not quite setup as necessary, then a race condition
style error can present itself when the application component starts up. This
change will add an 'oc get pods -w' to allow the admin to watch the pods as
they spin up. Pressing Ctrl+C will then get trapped and the watch can exit
and carry on adding the rest of the manifests to memory.

Ultimately this change is for current testing on 3.11. The process will change
shortly after the migration to OpenShift 4.x.