Fix testing errors
In this change an update to the SmartGateway object template is done
to match the changes done in the Smart Gateway Operator (change of
amqpmetrics_url to amqp_url).

Update the OCP setup script used within Travis-CI to match the latest
version of Ubuntu 16.04 (which now uses systemd and Docker 18.06-ce).

Closes #46
Closes #44