This adds the automated performance test to SAF for use with minishift... (#49)
This adds the automated performance test to SAF for use with minishift.

The performance test deploys two pods that execute test logic and host Grafana for data analysis. Collectd-tg is used to pump simulated network data through the smart gateway and into the prometheus database at user specified rates, intervals, and sizes. Grafana dashboards are automatically generated to graphically display queries to the database during the time of each test.