Add events testing to smoke test
Add functionality to the smoketest_entrypoint.sh that will return the
number of documents created during the test. Testing will return only
true if both the events and metrics smoke tests return without issue.

Updates the collectd minimal configuration for smoke testing to load
the appropriate plugins so that the threshold plugin will result in a
bunch of events being written into an index within ElasticSearch.