Use tls_server_name for ElasticSearch connection
When connecting the Smart Gateway (events type) to ElasticSearch,
there is a TlsServerName configuration we can pass which allows for a
better TLS connection setup (rather than just ignoring the names used
in the certificates). By default the hostname is going to be
elasticsearch.sa-telemetry.svc which is the internal DNS name loaded
into the certificates name alias list.

Set tls_server_name=elasticsearch.sa-telemetry.svc as the default name
when connecting and verifying the name at client connection time.