Update quickstart.sh for new application registry (#111)
We changed the name and path of the application and container registries in order
to avoid named registries. This change updates the quickstart.sh to account for this
change so that a Service Assurance deployment results.

Name of the Subscription has been updated to reflect the pattern that OLM created
for other dependent Subscriptions of the SAO. Updated spec.name to account for the
new package name.