Provide way of passing configs to quickstart.sh
After I tried to add a third quickstart-style script I decided it was
time to allow ability to pass in configs from a file so that we could
create whatever scenarios we desire for testing.

Sets quickstart.sh to use the existing 'default' configuration
previously. Adds both the ES Workaround and new Quicklab configuration
to a configs/ directory.